requirements.txt
python-dotenv
langchain
openai
pinecone-client
unstructured
tabulate
pdf2image
streamlit
tiktoken
